myapp/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
gunicorn==20.1.0
configparser==5.3.0